# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditActor": {
        "additionalProperties": false,
        "description": "Who did it. The email rather than the id alone: an audit nobody can read is a log.",
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email"
        ],
        "title": "AuditActor",
        "type": "object"
      },
      "AuditEntry": {
        "additionalProperties": false,
        "description": "One row of the audit.\n\n`ipHash` is stored and never sent. A twelve-character HMAC on screen tells a reader nothing\nand is evidence leaving the machine that holds the pepper; the column is there for the day an\ninvestigation asks the database, not for a column on a table.",
        "properties": {
          "action": {
            "enum": [
              "subscription.subscribe",
              "subscription.cancel",
              "subscription.change_plan",
              "subscription.redeem",
              "subscription.payment",
              "subscription.assign_program",
              "role.create",
              "role.update",
              "role.delete"
            ],
            "title": "Action",
            "type": "string"
          },
          "actor": {
            "$ref": "#/components/schemas/AuditActor"
          },
          "errorCode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "occurredAt": {
            "format": "date-time",
            "title": "Occurredat",
            "type": "string"
          },
          "outcome": {
            "enum": [
              "ok",
              "refused"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "targetId": {
            "title": "Targetid",
            "type": "string"
          },
          "targetType": {
            "enum": [
              "subscription",
              "role"
            ],
            "title": "Targettype",
            "type": "string"
          },
          "worldId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worldid"
          }
        },
        "required": [
          "id",
          "occurredAt",
          "actor",
          "action",
          "targetType",
          "targetId",
          "outcome",
          "payload"
        ],
        "title": "AuditEntry",
        "type": "object"
      },
      "AuditPage": {
        "additionalProperties": false,
        "description": "GET /api/audit.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditEntry"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "pageSize": {
            "title": "Pagesize",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ],
        "title": "AuditPage",
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "The contents of the envelope.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorCode": {
        "description": "The closed catalogue of machine-readable failures.\n\nClosed is the point: the frontend has a switch over these values, and a code invented at a\ncall site is a code that falls through it.",
        "enum": [
          "INVALID_CREDENTIALS",
          "NOT_AUTHENTICATED",
          "TOKEN_EXPIRED",
          "USER_INACTIVE",
          "PERMISSION_DENIED",
          "REFRESH_TOKEN_INVALID",
          "REFRESH_TOKEN_REUSED",
          "RATE_LIMITED",
          "VALIDATION_ERROR",
          "NOT_FOUND",
          "METHOD_NOT_ALLOWED",
          "INTERNAL_ERROR",
          "ROLE_IS_SYSTEM",
          "ROLE_IN_USE",
          "ROLE_CODE_TAKEN",
          "ALREADY_SUBSCRIBED",
          "UNKNOWN_PLAN",
          "UNKNOWN_PROMO_CODE",
          "PROMO_LIMIT_REACHED",
          "PROMO_ALREADY_BOUND",
          "UNKNOWN_REFERRAL_PROGRAM"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorEnvelope": {
        "description": "The body of every non-2xx response this service produces.\n\nRoutes declare it in `responses=` so the published schema says so as well.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "SubscriberPage": {
        "additionalProperties": false,
        "description": "GET /api/subscribers. The shape the specification fixes: items, total, page, pageSize.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SubscriberSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "pageSize": {
            "title": "Pagesize",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ],
        "title": "SubscriberPage",
        "type": "object"
      },
      "SubscriberSummary": {
        "additionalProperties": false,
        "description": "One row of the subscriber table.\n\nThe date fields are a discriminated view rather than a bag of optionals: `trialEndsAt` is set\nonly in TRIAL, `graceEndsAt` only in GRACE. The frontend's tagged union is built on that, so a\nfield that is present in a state it does not belong to would be a type that lies.\n\n`accessUntil` sits alongside them rather than replacing them. It is the one of the three that\nis true in the state the row is in, which is what a table with a single date column has to\nshow; the three stay because they are different questions and the subscriber card asks all of\nthem separately.",
        "properties": {
          "accessUntil": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accessuntil"
          },
          "cancelledAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelledat"
          },
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "expiresAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiresat"
          },
          "graceEndsAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graceendsat"
          },
          "lastActiveAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastactiveat"
          },
          "pendingPlanId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pendingplanid"
          },
          "planId": {
            "title": "Planid",
            "type": "string"
          },
          "promoCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Promocode"
          },
          "referrerId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Referrerid"
          },
          "state": {
            "enum": [
              "trial",
              "active",
              "grace",
              "expired",
              "cancelled"
            ],
            "title": "State",
            "type": "string"
          },
          "trialEndsAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trialendsat"
          },
          "userId": {
            "title": "Userid",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "displayName",
          "state",
          "planId"
        ],
        "title": "SubscriberSummary",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "substate-admin API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/audit": {
      "get": {
        "operationId": "list_page_api_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 1000000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Pagesize",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "actorUserId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actoruserid"
            }
          },
          {
            "in": "query",
            "name": "action",
            "required": false,
            "schema": {
              "items": {
                "enum": [
                  "subscription.subscribe",
                  "subscription.cancel",
                  "subscription.change_plan",
                  "subscription.redeem",
                  "subscription.payment",
                  "subscription.assign_program",
                  "role.create",
                  "role.update",
                  "role.delete"
                ],
                "type": "string"
              },
              "title": "Action",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "targetId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Targetid"
            }
          },
          {
            "in": "query",
            "name": "outcome",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ok",
                    "refused"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Outcome"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "One page of what operators did, newest first",
        "tags": [
          "audit"
        ]
      }
    },
...
    "/api/subscribers": {
      "get": {
        "operationId": "list_page_api_subscribers_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 1000000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Pagesize",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "-lastActiveAt",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "items": {
                "enum": [
                  "trial",
                  "active",
                  "grace",
                  "expired",
                  "cancelled"
                ],
                "type": "string"
              },
              "title": "State",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "planId",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Planid",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "cohort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "quiet",
                    "trial-ending",
                    "cancelled-still-active"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cohort"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriberPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "One page of subscribers, filtered and sorted",
        "tags": [
          "subscribers"
        ]
      }
    },
...
  }
}